Refactor into RESTful architecture
